commit message

+++ routes/web.php
@@ -15,21 +15,20 @@ Route::get('repairs/searchDevices', [RepairController::class, 'searchDevices'])-
 
 Route::controller(ServiceController::class)->group(function () {
     Route::get('/RepairVault', 'index')->name('RepairVault.main');
-    Route::get('/RepairVault/create',  'create')->name('RepairVault.create');
-    Route::post('/RepairVault',  'store')->name('RepairVault.store');
-    Route::get('/RepairVault/{service}', 'edit')->name('RepairVault.edit');
-    Route::put('/RepairVault/{service}','update')->name('RepairVault.update');
-    Route::delete('/RepairVault/{service}',  'destroy')->name('RepairVault.destroy');
+    Route::get('/RepairVault/create',  'create')->name('RepairVault.create')->middleware('can:is-admin');
+    Route::post('/RepairVault',  'store')->name('RepairVault.store')->middleware('can:is-admin');
+    Route::get('/RepairVault/{service}', 'edit')->name('RepairVault.edit')->middleware('can:is-admin');
+    Route::put('/RepairVault/{service}','update')->name('RepairVault.update')->middleware('can:is-admin');
+    Route::delete('/RepairVault/{service}',  'destroy')->name('RepairVault.destroy')->middleware('can:is-admin');
     Route::get('/search', 'search')->name('RepairVault.search');
-
 });
 
 Route::controller(RepairController::class)->group(function () {
-    Route::get('/repairs/create',  'create')->name('repairs.create');
-    Route::post('/repairs',  'store')->name('repairs.store');
-    Route::get('/repairs/{repair}',  'edit')->name('repairs.edit');
-    Route::put('/repairs/{repair}',  'update')->name('repairs.update');
-    Route::delete('/repairs/{repair}',  'destroy')->name('repairs.destroy');
+    Route::get('/repairs/create',  'create')->name('repairs.create')->middleware('can:is-admin');
+    Route::post('/repairs',  'store')->name('repairs.store')->middleware('can:is-admin');
+    Route::get('/repairs/{repair}',  'edit')->name('repairs.edit')->middleware('can:is-admin');
+    Route::put('/repairs/{repair}',  'update')->name('repairs.update')->middleware('can:is-admin');
+    Route::delete('/repairs/{repair}',  'destroy')->name('repairs.destroy')->middleware('can:is-admin');
 });
 
 Route::controller(AuthController::class)->group(function () {
@@ -37,4 +36,3 @@ Route::controller(AuthController::class)->group(function () {
     Route::post('/auth/login', 'authenticate')->name('login.authenticate');
     Route::get('/auth/logout', 'logout')->name('logout');
 });
-
